[sys-manage] `install` is no longer experimental!
Note, however, that I still haven't included a Brewfile in this repo.
One must be manually placed at ~/.Brewfile for the command to do its
job. I haven't decided if I'm going to change this or not, leaving it
out forces me to reconsider my installed software once in a while.