[zsh] Remove battery, optimize prompt a bit, clean
I've disabled the _approximate completer for the time being, because it
often ends up getting me further from what I was trying to do when I've
missed too many keys. Might work on fixing it up later.

Also fixed it so that the tmux window will be named when the shell first
opens, instead of just after changing directories. This plays nicer with
having tmux set up to land you in the same directory as the last pane
you were in.