[profile] Expose ssh-agent to the whole session
By running keychain in .profile, SDDM will execute the agent before
starting the compositor, meaning that tools started outside a shell have
access to it. The first shell to be spawned within the session is still
responsible for actually adding keys, as there is no way to prompt from
.profile.

keychain no longer tries to manage the GPG agent, as there is a systemd
user unit handling this.