[scripts/bar] Add UPower-based battery block
Because my system has multiple batteries, acpi gives a completely
inaccurate estimate for both time and percentage remaining. Use UPower
to monitor battery state and calculate it ourselves.

This also provides a place to do some smoothing of the estimated time
remaining, but I'm saving that for a rainy-day project.