[sway+scripts] Add sway-loner
This is a little ipc daemon that monitors the workspace layout. When it
determines that only a single top-level container is visible, it changes
the workspace's horizontal gaps to make that container to 1024 pixels
wide.

The primary non-terminal content in my workspaces is qutebrowser, and
I've noticed that 1024 is the upper limit of useful content for most
websites. Shrinking the browser down to match that wraps text at a nicer
width and also lets me see more of oguri rather than dead space in the
site layout.

There's an ignore list which currently just contains mpv, so when
I watch video it will take up the entire workspace.

When the workspace splits into multiple top-level containers, the full
width is used again. I rather like the effect.